+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.2.0
 * Update Company API
 * Create Tow API
+* Create Metric API
 
 ## 0.1.1
 * Add logs to user handler method